Revert "Merge pull request #35148 from ifreund/player-to-character"
This reverts commit 0278ae1ce7112bc4b7a2583b84a515ea6c15cd7f.
