+++ src/savegame_json.cpp
@@ -478,19 +478,6 @@ void Character::load( JsonObject &data )
     data.read( "healthy_mod", healthy_mod );
     data.read( "healed_24h", healed_total );
 
-    // status
-    temp_cur.fill( 5000 );
-    data.read( "temp_cur", temp_cur );
-
-    temp_conv.fill( 5000 );
-    data.read( "temp_conv", temp_conv );
-
-    frostbite_timer.fill( 0 );
-    data.read( "frostbite_timer", frostbite_timer );
-
-    body_wetness.fill( 0 );
-    data.read( "body_wetness", body_wetness );
-
     //energy
     data.read( "stim", stim );
     data.read( "stamina", stamina );
@@ -674,12 +661,6 @@ void Character::store( JsonOut &json ) const
     json.member( "healthy_mod", healthy_mod );
     json.member( "healed_24h", healed_total );
 
-    // status
-    json.member( "temp_cur", temp_cur );
-    json.member( "temp_conv", temp_conv );
-    json.member( "frostbite_timer", frostbite_timer );
-    json.member( "body_wetness", body_wetness );
-
     // needs
     json.member( "thirst", thirst );
     json.member( "hunger", hunger );
@@ -773,6 +754,7 @@ void player::store( JsonOut &json ) const
     json.member( "reactor_plut", reactor_plut );
     json.member( "slow_rad", slow_rad );
     json.member( "scent", static_cast<int>( scent ) );
+    json.member( "body_wetness", body_wetness );
 
     // gender
     json.member( "male", male );
@@ -999,10 +981,6 @@ void avatar::store( JsonOut &json ) const
     json.member( "temp_conv", temp_conv );
     json.member( "frostbite_timer", frostbite_timer );
 
-    // crafting etc
-    json.member( "activity", activity );
-    json.member( "backlog", backlog );
-
     // npc: unimplemented, potentially useful
     json.member( "learned_recipes", *learned_recipes );
 
@@ -1103,6 +1081,17 @@ void avatar::load( JsonObject &data )
         }
         g->scen = generic_scenario;
     }
+    temp_cur.fill( 5000 );
+    data.read( "temp_cur", temp_cur );
+
+    temp_conv.fill( 5000 );
+    data.read( "temp_conv", temp_conv );
+
+    frostbite_timer.fill( 0 );
+    data.read( "frostbite_timer", frostbite_timer );
+
+    body_wetness.fill( 0 );
+    data.read( "body_wetness", body_wetness );
 
     data.read( "learned_recipes", *learned_recipes );
     valid_autolearn_skills->clear(); // Invalidates the cache
